requirements.txt
amqp==5.3.1
annotated-types==0.7.0
asgiref==3.8.1
attrs==25.3.0
billiard==4.2.1
cachetools==6.0.0
celery==5.5.3
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cron-descriptor==1.4.5
Django==5.2.2
django-celery==3.1.17
django-celery-beat==2.8.1
django-jazzmin==3.0.1
django-jsonform==2.23.2
django-rql==4.4.1
django-timezone-field==7.1
django_celery_results==2.6.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
drf-spectacular==0.28.0
idna==3.10
inflection==0.5.1
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
kombu==5.5.4
lark-parser==0.11.0
lib-rql==2.0.2
packaging==25.0
prompt_toolkit==3.0.51
psycopg2-binary==2.9.10
pydantic==2.11.5
pydantic_core==2.33.2
PyJWT==2.9.0
python-crontab==3.2.0
python-dateutil==2.9.0.post0
python-decouple==3.8
PyYAML==6.0.2
referencing==0.36.2
requests==2.32.3
rpds-py==0.25.1
six==1.17.0
sqlparse==0.5.3
typing-inspection==0.4.1
typing_extensions==4.14.0
tzdata==2025.2
ulid-py==1.1.0
uritemplate==4.2.0
urllib3==2.4.0
vine==5.1.0
wcwidth==0.2.13
